move icon-types to utils folder
