+++ src/components/common/icons/utils/icon-types.ts
